Moved IsGlobal domain logic
