+++ src/Equinor.ProCoSys.Completion.Query/MailTemplateQueries/GetAllMailTemplates/MailTemplateDto.cs
@@ -1,8 +1,3 @@
-﻿using Equinor.ProCoSys.Common.Misc;
+﻿namespace Equinor.ProCoSys.Completion.Query.MailTemplateQueries.GetAllMailTemplates;
 
-namespace Equinor.ProCoSys.Completion.Query.MailTemplateQueries.GetAllMailTemplates;
-
-public record MailTemplateDto(string Code, string Subject, string Body, bool IsVoided, string? Plant)
-{
-    public bool IsGlobal => Plant.IsEmpty();
-}
+public record MailTemplateDto(string Code, string Subject, string Body, bool IsVoided, string? Plant, bool IsGlobal);
